refactor: AI 지출 분석 기능 분리 (백엔드 & 프론트엔드)
