+++ src/main/java/store/lastdance/dto/analysis/AnalyzeExpenseRequestDTO.java
@@ -1,0 +1,10 @@
+package store.lastdance.dto.analysis;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import java.time.LocalDate;
+
+@Schema(description = "LLM 분석 요청")
+public record AnalyzeExpenseRequestDTO(
+        LocalDate startDate,
+        LocalDate endDate
+) {}
