+++ src/main/java/store/lastdance/dto/analysis/ExpenseAnalysisHistoryDTO.java
@@ -1,4 +1,4 @@
-package store.lastdance.dto.expense;
+package store.lastdance.dto.analysis;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
