+++ src/main/java/store/lastdance/repository/analysis/ExpenseAnalysisHistoryRepository.java
@@ -1,4 +1,4 @@
-package store.lastdance.repository.expense;
+package store.lastdance.repository.analysis;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -8,14 +8,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import store.lastdance.domain.expense.ExpenseAnalysisHistory;
+import store.lastdance.domain.analysis.ExpenseAnalysisHistory;
 import store.lastdance.domain.user.User;
 import store.lastdance.dto.admin.AdminExpenseAnalyzerHistoryDTO;
 import store.lastdance.dto.admin.ExpenseAnalyzerFeedbackStatsDTO;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.UUID;
 
 public interface ExpenseAnalysisHistoryRepository extends JpaRepository<ExpenseAnalysisHistory, Long>,
         JpaSpecificationExecutor<ExpenseAnalysisHistory> {
