+++ src/main/java/store/lastdance/service/admin/AdminServiceImpl.java
@@ -15,7 +15,7 @@ import store.lastdance.domain.admin.ReportType;
 import store.lastdance.domain.aijudgment.AiJudgment;
 import store.lastdance.domain.community.Comment;
 import store.lastdance.domain.community.Post;
-import store.lastdance.domain.expense.ExpenseAnalysisHistory;
+import store.lastdance.domain.analysis.ExpenseAnalysisHistory;
 import store.lastdance.domain.user.User;
 import store.lastdance.domain.user.UserRole;
 import store.lastdance.domain.user.OAuthProvider;
@@ -25,7 +25,7 @@ import store.lastdance.dto.admin.stats.DashboardUserStats;
 import store.lastdance.dto.admin.stats.UserStats;
 import store.lastdance.dto.expense.ExpenseAnalysisHistoryDTO;
 import store.lastdance.repository.aijudgment.AiJudgmentRepository;
-import store.lastdance.repository.expense.ExpenseAnalysisHistoryRepository;
+import store.lastdance.repository.analysis.ExpenseAnalysisHistoryRepository;
 import store.lastdance.repository.user.UserRepository;
 import store.lastdance.repository.community.PostRepository;
 import store.lastdance.repository.community.CommentRepository;
