--- src/main/java/store/lastdance/service/expense/ExpenseAnalyzer.java
@@ -1,7 +1,0 @@
-package store.lastdance.service.expense;
-
-import store.lastdance.dto.expense.AnalyzeExpenseResponseDTO;
-
-public interface ExpenseAnalyzer {
-    AnalyzeExpenseResponseDTO.Suggestion analyzerExpenseData(String expenseJson);
-}
